refactor: BaseEntity 디렉토리 경로 수정
- global.common.domain.BaseEntity.class 경로 수정

#11
